fix Magic Missile Spell breaking multi blocks

+++ src/main/java/com/axanthic/icaria/common/entity/MagicMissileSpellEntity.java
@@ -1,9 +1,13 @@
 package com.axanthic.icaria.common.entity;
 
+import com.axanthic.icaria.common.block.KettleBlock;
+import com.axanthic.icaria.common.block.TripleBarrelRackBlock;
+
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.DoorBlock;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.EntityHitResult;
 
@@ -22,7 +26,9 @@ public class MagicMissileSpellEntity extends SpellEntity {
 		super.onHitBlock(pResult);
 		var level = this.level();
 		var blockPos = pResult.getBlockPos();
-		if (level.getBlockState(blockPos).getDestroySpeed(level, blockPos) > 0) {
+		var blockState = level.getBlockState(blockPos);
+		var block = blockState.getBlock();
+		if (level.getBlockState(blockPos).getDestroySpeed(level, blockPos) > 0 && level.getBlockEntity(blockPos) == null && !(block instanceof DoorBlock) && !(block instanceof KettleBlock) && !(block instanceof TripleBarrelRackBlock)) {
 			level.destroyBlock(blockPos, false);
 		}
 	}
